Do not run toWorldCoords if unneccessary, add testcases
Since both radiusArc and sagittaArc boils down to a threePointArc in the end I want to be able to just return a threePointArc.
This increases code reuse. However, the function toWorldCoords seems to do something strange to a vector that already is in World coordinates.
This commit checks if the point already is an Vector before trying to make it an vector.

I also refined the testcases by checking edge length and edge type.